Update 990_verify_rootfs.sh
Let grep match to the end of the line
to avoid that falsely a prefix matches
e.g. when not_found_library="libparted.so.2"
does not exist but libparted.so.2.0.1 exists.
Always use 'find' with its -xdev option to be on the safe side
against possibly mounted things within the recovery system.