In 990_verify_rootfs.sh show relative paths in user messages for files in ROOTFS_DIR (#3308)
In build/default/990_verify_rootfs.sh
enhanced how files from within the ReaR recovery system
(i.e. files from within ROOTFS_DIR) are shown to the user:
In user messages it is misleading to show the full path from
within the recovery system without leading ROOTFS_DIR
for example showing /bin/parted for what actually is
/var/tmp/rear.XXX/rootfs/bin/parted
because for the user /bin/parted means the full path
on his original system but there is no /bin/parted on
the original system because there it is /usr/sbin/parted
so to show files inside the recovery system to the user
even without (the long) leading ROOTFS_DIR we can show them
as relative path i.e. without leading slash e.g. as bin/parted
when from the context it should be clear enough
that a file inside the recovery system is meant.