Create /var/run symlink dynamically during mkrescue
Introduce a script that can create symlinks and empty directories in the
rescue ramdisk if they exist in original system. This way one can avoid
having plenty of empty directories under usr/share/rear/skel.

For now, use it to create the /var/run symlink. Not having it as an
actual symlink in the source tree helps avoiding pitfalls associated
with symlinks in the source tree, in particular the packaging problem
that RPM is not able to change a directory into a symlink and vice-versa
during upgrade (before ReaR 2.8 it was a directory, not a symlink):
https://docs.fedoraproject.org/en-US/packaging-guidelines/Directory_Replacement/
and issue #3375.