Linux-ppc64le/660_install_grub2: fix compatibility with PowerNV
Modern PowerNV systems with the OPAL firmware do not require the presence of
the PPC PReP boot partition [1,2].  Therefore, the installation of GRUB2
should be made optional on these machines if such partition does not exists.

As long as the GRUB2 configuration is created successfully, Petitboot will be
able to load it and boot the restored system.  That was already the case, this
commit just fixes the annoying warning message about a missing bootloader.

[1] https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/8/html-single/automatically_installing_rhel/index#raid_kickstart-commands-for-handling-storage
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1970432#c2