Delete usr/share/rear/build/USB directory (#3435)
Delete usr/share/rear/build/USB/default/800_enforce_usb_output.sh
which is the only file in the usr/share/rear/build/USB directory
so this whole directory gets deleted.
For the reasons why 800_enforce_usb_output.sh is deleted
see the comments in this file e.g. in ReaR 2.9
https://github.com/rear/rear/blob/rear-2.9/usr/share/rear/build/USB/default/800_enforce_usb_output.sh
and see the recent
https://github.com/rear/rear/pull/3424#discussion_r2009556786
and the former discussion in
https://github.com/rear/rear/pull/3103