feat(rear): exclude hard disks with COVE_RESCUE label
Since Cove Rescue Media is VHD disk in Azure env, it is considered a
hard disk by ReaR and, as a result, might be used as a possible target
for disks mapping. To avoid this, we are excluding the rescue system
by the COVE_RESCUE label.

Jira-Ref: BCF-5849: Support Bare Metal Recovery in Azure